+++ src/main/java/it/gov/pagopa/pu/debtpositions/service/installmentsync/BaseInstallmentSynchronizeService.java
@@ -2,8 +2,22 @@ package it.gov.pagopa.pu.debtpositions.service.installmentsync;
 
 import it.gov.pagopa.pu.debtpositions.dto.generated.DebtPositionDTO;
 import it.gov.pagopa.pu.debtpositions.dto.generated.InstallmentDTO;
+import it.gov.pagopa.pu.debtpositions.exception.custom.NotFoundException;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
 
-public interface BaseInstallmentSynchronizeService {
+@Service
+@Slf4j
+public abstract class BaseInstallmentSynchronizeService{
 
-  InstallmentDTO findInstallment(DebtPositionDTO debtPositionDTO, String iud, Integer paymentOptionIndex);
+  public InstallmentDTO findInstallment(DebtPositionDTO debtPositionDTO, String iud, Integer paymentOptionIndex) {
+    return debtPositionDTO.getPaymentOptions().stream()
+      .filter(po -> po.getPaymentOptionIndex().equals(paymentOptionIndex))
+      .flatMap(po -> po.getInstallments().stream())
+      .filter(inst -> inst.getIud().equals(iud))
+      .findFirst()
+      .orElseThrow(() -> new NotFoundException(
+        String.format("The installment with iud %s in payment option with index %s not found", iud, paymentOptionIndex)));
+  }
+
 }
