+++ src/test/java/it/gov/pagopa/pu/debtpositions/service/installmentsync/operation/InstallmentSynchronizeCancelServiceImplTest.java
@@ -3,7 +3,6 @@ package it.gov.pagopa.pu.debtpositions.service.installmentsync.operation;
 import it.gov.pagopa.pu.debtpositions.dto.generated.*;
 import it.gov.pagopa.pu.debtpositions.exception.custom.ConflictErrorException;
 import it.gov.pagopa.pu.debtpositions.exception.custom.NotFoundException;
-import it.gov.pagopa.pu.debtpositions.service.installmentsync.BaseInstallmentSynchronizeService;
 import it.gov.pagopa.pu.debtpositions.service.update.DebtPositionCancelInstallmentService;
 import org.apache.commons.lang3.tuple.Pair;
 import org.junit.jupiter.api.BeforeEach;
@@ -26,14 +25,12 @@ class InstallmentSynchronizeCancelServiceImplTest {
 
   @Mock
   private DebtPositionCancelInstallmentService debtPositionCancelInstallmentServiceMock;
-  @Mock
-  private BaseInstallmentSynchronizeService baseInstallmentSynchronizeServiceMock;
 
   private InstallmentSynchronizeCancelServiceImpl installmentSynchronizeCancelService;
 
   @BeforeEach
   void setUp() {
-    installmentSynchronizeCancelService = new InstallmentSynchronizeCancelServiceImpl(debtPositionCancelInstallmentServiceMock, baseInstallmentSynchronizeServiceMock);
+    installmentSynchronizeCancelService = new InstallmentSynchronizeCancelServiceImpl(debtPositionCancelInstallmentServiceMock);
   }
 
   @Test
@@ -43,11 +40,11 @@ class InstallmentSynchronizeCancelServiceImplTest {
     String operatorExternalUserId = "operatorExternalUserId";
     String workflowId = "workflowId";
     InstallmentSynchronizeDTO installmentSynchronizeDTO = buildInstallmentSynchronizeDTO();
+    installmentSynchronizeDTO.setIud("iud");
+    installmentSynchronizeDTO.setPaymentOptionIndex(1);
     DebtPositionDTO debtPositionDTO = buildDebtPositionDTO();
     InstallmentDTO installmentDTO = buildInstallmentDTO();
 
-    Mockito.when(baseInstallmentSynchronizeServiceMock.findInstallment(debtPositionDTO, installmentSynchronizeDTO.getIud(), installmentSynchronizeDTO.getPaymentOptionIndex()))
-      .thenReturn(installmentDTO);
     Mockito.when(debtPositionCancelInstallmentServiceMock.cancelInstallment(debtPositionDTO, List.of(installmentDTO), massive, accessToken, operatorExternalUserId))
       .thenReturn(Pair.of(debtPositionDTO, workflowId));
 
@@ -68,7 +65,21 @@ class InstallmentSynchronizeCancelServiceImplTest {
     assertEquals(String.format("The debt position related to iupd %s was not found", installmentSynchronizeDTO.getIupdOrg()), notFoundException.getMessage());
   }
 
+  @Test
+  void testInstallmentSyncCancelWithInstallmentNotFoundThenThrowException() {
+    boolean massive = true;
+    String accessToken = "accessToken";
+    String operatorExternalUserId = "operatorExternalUserId";
+    InstallmentSynchronizeDTO installmentSynchronizeDTO = buildInstallmentSynchronizeDTO();
+    installmentSynchronizeDTO.setIud("iud2");
+    installmentSynchronizeDTO.setPaymentOptionIndex(2);
+    DebtPositionDTO debtPositionDTO = buildDebtPositionDTO();
 
+    NotFoundException notFoundException = assertThrows(NotFoundException.class, () ->
+      installmentSynchronizeCancelService.syncInstallment(installmentSynchronizeDTO, debtPositionDTO, massive, accessToken, operatorExternalUserId));
+    assertEquals(String.format("The installment with iud %s in payment option with index %s not found",
+      installmentSynchronizeDTO.getIud(), installmentSynchronizeDTO.getPaymentOptionIndex()), notFoundException.getMessage());
+  }
 
   @Test
   void testInstallmentSyncCancelWithInstallmentPaidThenThrowException() {
@@ -76,11 +87,10 @@ class InstallmentSynchronizeCancelServiceImplTest {
     String accessToken = "accessToken";
     String operatorExternalUserId = "operatorExternalUserId";
     InstallmentSynchronizeDTO installmentSynchronizeDTO = buildInstallmentSynchronizeDTO();
+    installmentSynchronizeDTO.setIud("iud");
+    installmentSynchronizeDTO.setPaymentOptionIndex(1);
     DebtPositionDTO debtPositionDTO = buildDebtPositionDTO();
     debtPositionDTO.getPaymentOptions().getFirst().getInstallments().getFirst().setStatus(InstallmentStatus.PAID);
-
-    Mockito.when(baseInstallmentSynchronizeServiceMock.findInstallment(debtPositionDTO, installmentSynchronizeDTO.getIud(), installmentSynchronizeDTO.getPaymentOptionIndex()))
-      .thenReturn(debtPositionDTO.getPaymentOptions().getFirst().getInstallments().getFirst());
 
     ConflictErrorException conflictException = assertThrows(ConflictErrorException.class, () ->
       installmentSynchronizeCancelService.syncInstallment(installmentSynchronizeDTO, debtPositionDTO, massive, accessToken, operatorExternalUserId));
@@ -94,12 +104,11 @@ class InstallmentSynchronizeCancelServiceImplTest {
     String accessToken = "accessToken";
     String operatorExternalUserId = "operatorExternalUserId";
     InstallmentSynchronizeDTO installmentSynchronizeDTO = buildInstallmentSynchronizeDTO();
+    installmentSynchronizeDTO.setIud("iud");
+    installmentSynchronizeDTO.setPaymentOptionIndex(1);
     installmentSynchronizeDTO.setIngestionFlowFileId(3L);
     DebtPositionDTO debtPositionDTO = buildDebtPositionDTO();
     debtPositionDTO.getPaymentOptions().getFirst().getInstallments().getFirst().setStatus(InstallmentStatus.TO_SYNC);
-
-    Mockito.when(baseInstallmentSynchronizeServiceMock.findInstallment(debtPositionDTO, installmentSynchronizeDTO.getIud(), installmentSynchronizeDTO.getPaymentOptionIndex()))
-      .thenReturn(debtPositionDTO.getPaymentOptions().getFirst().getInstallments().getFirst());
 
     ConflictErrorException conflictException = assertThrows(ConflictErrorException.class, () ->
       installmentSynchronizeCancelService.syncInstallment(installmentSynchronizeDTO, debtPositionDTO, massive, accessToken, operatorExternalUserId));
@@ -121,9 +130,6 @@ class InstallmentSynchronizeCancelServiceImplTest {
     debtPositionDTO.getPaymentOptions().getFirst().getInstallments().getFirst().setSyncStatus(
       InstallmentSyncStatus.builder().syncStatusTo(InstallmentStatus.PAID).build());
 
-    Mockito.when(baseInstallmentSynchronizeServiceMock.findInstallment(debtPositionDTO, installmentSynchronizeDTO.getIud(), installmentSynchronizeDTO.getPaymentOptionIndex()))
-      .thenReturn(debtPositionDTO.getPaymentOptions().getFirst().getInstallments().getFirst());
-
     ConflictErrorException conflictException = assertThrows(ConflictErrorException.class, () ->
       installmentSynchronizeCancelService.syncInstallment(installmentSynchronizeDTO, debtPositionDTO, massive, accessToken, operatorExternalUserId));
     assertEquals(String.format("The installment with %s cannot be cancelled because is not in an allowed status to: %s",
@@ -144,9 +150,6 @@ class InstallmentSynchronizeCancelServiceImplTest {
     debtPositionDTO.getPaymentOptions().getFirst().getInstallments().getFirst().setStatus(InstallmentStatus.TO_SYNC);
     debtPositionDTO.getPaymentOptions().getFirst().getInstallments().getFirst().setSyncStatus(null);
 
-    Mockito.when(baseInstallmentSynchronizeServiceMock.findInstallment(debtPositionDTO, installmentSynchronizeDTO.getIud(), installmentSynchronizeDTO.getPaymentOptionIndex()))
-      .thenReturn(debtPositionDTO.getPaymentOptions().getFirst().getInstallments().getFirst());
-
     Mockito.when(debtPositionCancelInstallmentServiceMock.cancelInstallment(debtPositionDTO,
         List.of(debtPositionDTO.getPaymentOptions().getFirst().getInstallments().getFirst()), massive, accessToken, operatorExternalUserId))
       .thenReturn(Pair.of(debtPositionDTO, workflowId));
@@ -155,31 +158,4 @@ class InstallmentSynchronizeCancelServiceImplTest {
 
     assertEquals(workflowId, result);
   }
-  @Test
-  void testInstallmentSyncCancelWithInstallmentWithSyncStatusValidThenOk() {
-    boolean massive = true;
-    String accessToken = "accessToken";
-    String operatorExternalUserId = "operatorExternalUserId";
-    String workflowId = "workflowId";
-    InstallmentSynchronizeDTO installmentSynchronizeDTO = buildInstallmentSynchronizeDTO();
-    installmentSynchronizeDTO.setIud("iud");
-    installmentSynchronizeDTO.setPaymentOptionIndex(1);
-    installmentSynchronizeDTO.setIngestionFlowFileId(1L);
-    DebtPositionDTO debtPositionDTO = buildDebtPositionDTO();
-    debtPositionDTO.getPaymentOptions().getFirst().getInstallments().getFirst().setStatus(InstallmentStatus.TO_SYNC);
-    debtPositionDTO.getPaymentOptions().getFirst().getInstallments().getFirst()
-            .setSyncStatus(InstallmentSyncStatus.builder().syncStatusTo(InstallmentStatus.UNPAID).build());
-
-    Mockito.when(baseInstallmentSynchronizeServiceMock.findInstallment(debtPositionDTO, installmentSynchronizeDTO.getIud(), installmentSynchronizeDTO.getPaymentOptionIndex()))
-            .thenReturn(debtPositionDTO.getPaymentOptions().getFirst().getInstallments().getFirst());
-
-    Mockito.when(debtPositionCancelInstallmentServiceMock.cancelInstallment(debtPositionDTO,
-                    List.of(debtPositionDTO.getPaymentOptions().getFirst().getInstallments().getFirst()), massive, accessToken, operatorExternalUserId))
-            .thenReturn(Pair.of(debtPositionDTO, workflowId));
-
-    String result = installmentSynchronizeCancelService.syncInstallment(installmentSynchronizeDTO, debtPositionDTO, massive, accessToken, operatorExternalUserId);
-
-    assertEquals(workflowId, result);
-  }
-
 }
